pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nineball-ga4"
version = "0.1.0"
description = "GA4 data extraction and analysis for ecommerce demo site"
authors = [{name = "Calvin W", email = "calvinw@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9,<3.13"
dependencies = [
    "google-analytics-data>=0.18.0",
    "google-auth>=2.0.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88